==> analysis/2.correlation.R <==
#' ---
#' title: "A correlation analysis of clinical variables of TCGA-KIRC patients"
#' output: 
#'   html_document: 
#'     default
#'   github_document: 
#'     df_print: paged
#'     html_preview: FALSE
#'     keep_html: TRUE
#'   pdf_document:
#'     latex_engine: xelatex
#' knit: (function(inputFile, encoding) {
#'   rmarkdown::render(inputFile, encoding = encoding, output_format = "all") })     
#' ---
#' 
#' This project contains a pipeline of clinical analysis of the Cancer Genome Atlas Kidney Renal Clear Cell Carcinoma (TCGA-KIRC) data of patients, from [Genomic Data Commons Data Portal](https://portal.gdc.cancer.gov/exploration?filters=%7B%22op%22%3A%22and%22%2C%22content%22%3A%5B%7B%22op%22%3A%22in%22%2C%22content%22%3A%7B%22field%22%3A%22cases.project.project_id%22%2C%22value%22%3A%5B%22TCGA-KIRC%22%5D%7D%7D%5D%7D) and [cBioPortal](https://www.cbioportal.org/study/summary?id=kirp_tcga).
#' 
#' Previously, we presented [an exploratory preprocessing analysis](1.preprocessing.md). In this section, Chi-squared test is applied to compare two or more proportions of categorical variables and T-student test to compare the means of numeric ones regardind the levels of 'Overall_Survival_Status'. The Hypoteis test is performed and p-value indicates the strength of evidence in supportting the null hypothesis.
#' 
#' 

#' 
#' 
## ----message=FALSE, warning=FALSE, paged.print=FALSE--------------------------
# Set the packages of interest
packages = c("tidyverse","skimr","finalfit","rstatix", "ggpubr","GGally", "plotly")

# if a package is installed, it will be loaded
# otherwise, the missing package(s) will be installed and loaded
package.check <- lapply(packages, FUN = function(x) {
  if (!require(x, character.only = TRUE)) {
    install.packages(x, dependencies = TRUE)
    library(x, character.only = TRUE)
  }
})

suppressMessages(library("tidyverse"))
rm(packages)
setwd(".")

#' 
#' ## 1. Importing data
#' 
## ----message=FALSE, warning=FALSE, paged.print=FALSE--------------------------
kirc_clin <- read_csv("data/kirc_clin.csv")

#' 
#' ## 2. Taming data 
#' 
## -----------------------------------------------------------------------------
kirc_clin <- kirc_clin %>%
  mutate_if(is.character, as.factor) %>%
  mutate(patient_id = as.character(patient_id),
         age = as.integer(age),
         year_diagnose = as.integer(year_diagnose))

# check 
glimpse(kirc_clin) 

#' 
#' ## 3.The dependent variable
#' 
## -----------------------------------------------------------------------------
# Check the number of levels. If greater than 2, it thas to run a simple logistic regression presents only two levels (otherwise, it has to run a ordinal logistic regression)
table(kirc_clin$over_surv_stt, exclude = NULL) 

#' 
#' ## 4. Numeric variables vs. over_surv_stt
#' 
#' Correlation matrix - graphic visualization
#' 
## ----message=FALSE, warning=FALSE---------------------------------------------
cols_numeric <- kirc_clin %>% 
  select_if(is.numeric) %>%
  names

kirc_clin_numeric <- kirc_clin %>%
                      select(one_of(c(cols_numeric, "over_surv_stt")))  

levels(kirc_clin_numeric$over_surv_stt) <- c("D","L")

ggpairs(kirc_clin_numeric, columns = cols_numeric, 
        title="Correlation matrix",               
        mapping= aes(colour = over_surv_stt), 
        upper = list(combo = wrap("box_no_facet", alpha=0.1), 
                     continuous = wrap("cor", size = 2, alignPercent = 0.8)),
        lower = list(continuous = wrap("smooth", alpha = 0.3, size=0.2) )) +
        theme(panel.background = element_rect(color = "black", size=0.5, fill="white"),
          panel.grid.major = element_blank()) 
        

#' 
#' Run multiple T-tests on over_surv_stt
#' 
#' Transform the data into long format
#' 
## -----------------------------------------------------------------------------
# Put all variables in the same column except `over_surv_stt`, the grouping variable

levels(kirc_clin_numeric$over_surv_stt) <- c("DECEASED","LIVING")

# Convert to Tidyverse
kirc_clin_numeric.long <- kirc_clin_numeric %>%
  pivot_longer(-over_surv_stt, names_to = "variables", values_to = "value")
kirc_clin_numeric.long <- kirc_clin_numeric.long[!is.na(kirc_clin_numeric.long$value), ]
kirc_clin_numeric.long$value.log <- log2(kirc_clin_numeric.long$value+1)

# OR
# kirc_clin_numeric.long <- kirc_clin_numeric %>% 
#   gather(key = 'variables', value = 'value', -over_surv_stt, na.rm = TRUE) %>%
#     mutate(value.log = log2(kirc_clin_numeric.long$value+1))

kirc_clin_numeric.long %>% sample_n(6) %>% knitr::kable(.)

#' 
#' Group the data by variables and compare over_surv_stt groups
#' 
#' Adjust the p-values and add significance levels
#' 
## -----------------------------------------------------------------------------
stat.test <- kirc_clin_numeric.long %>%
  group_by(variables) %>%
  t_test(value ~ over_surv_stt) %>%
  adjust_pvalue(method = "BH") %>%
  add_significance()
stat.test %>% knitr::kable(.)

#' 
#' 
## -----------------------------------------------------------------------------
# Create the plot on logscale
myplot <- ggboxplot(
  kirc_clin_numeric.long, x = "over_surv_stt", y = "value.log",
  fill = "over_surv_stt", palette = "npg", legend = "none", 
  ggtheme = theme_pubr(border = TRUE)
  ) +
  facet_wrap(~variables)

# Add statistical test p-values
# OBS: different p-values over vaule vs. log.value!! 
stat.test <- stat.test %>% add_xy_position(x = "over_surv_stt")
myplot + stat_pvalue_manual(stat.test, label = "p.adj.signif")

#' 
#' 
## -----------------------------------------------------------------------------
# Group the data by variables and do a graph for each variable
graphs <- kirc_clin_numeric.long %>%
  group_by(variables) %>%
  doo(
    ~ggboxplot(
      data =., x = "over_surv_stt", y = "value",
      fill = "over_surv_stt", palette = "npg", legend = "none",
      ggtheme = theme_pubr()
      )  +
      geom_jitter(width = 0.05, alpha = 0.2, color = "orange"), 
    result = "plots"
  )
graphs %>% knitr::kable(.)

#' 
#' 
## -----------------------------------------------------------------------------
# Add statitistical tests to each corresponding plot
# variables <- graphs$variables
# for(i in 1:length(variables)){
#   graph.i <- graphs$plots[[i]] + 
#     labs(title = variables[i]) +
#     #stat_pvalue_manual(stat.test[i, ], label = "p.adj.signif")
#   print(graph.i)
# }
# Error in print(graph.i) : objeto 'graph.i' não encontrado

#' 
#' 
## -----------------------------------------------------------------------------
# ggplot(kirc_clin, aes(age, fill= over_surv_stt)) +
#   geom_histogram(bins = 15, position = "dodge")
# t.test(kirc_clin$age ~ kirc_clin$over_surv_stt) 
# 
# ggplot(kirc_clin, aes(year_diagnose, fill= over_surv_stt)) +
#   geom_histogram(bins = 15, position = "dodge")
# t.test(kirc_clin$year_diagnose ~ kirc_clin$over_surv_stt) 
# 
# ggplot(kirc_clin, aes(x=over_surv_stt, y=disease_free_mth)) +
#   geom_boxplot(width = .5) +
#   geom_jitter(width = 0.05, alpha = 0.2, color = "orange")
# t.test(kirc_clin$disease_free_mth ~ kirc_clin$over_surv_stt)
# 
# ggplot(kirc_clin, aes(x=over_surv_stt, y=frac_genome_alter)) +
#   geom_boxplot(width = .5) +
#   geom_jitter(width = 0.05, alpha = 0.2, color = "orange")
# t.test(kirc_clin$frac_genome_alter ~ kirc_clin$over_surv_stt)
# 
# ggplot(kirc_clin, aes(x=over_surv_stt, y=long_dim)) +
#   geom_boxplot(width = .5) +
#   geom_jitter(width = 0.05, alpha = 0.2, color = "orange")
# t.test(kirc_clin$long_dim ~ kirc_clin$over_surv_stt)
# 
# ggplot(kirc_clin, aes(x=over_surv_stt, y=mutation_cnt)) +
#   geom_boxplot(width = .5) +
#   geom_jitter(width = 0.05, alpha = 0.2, color = "orange")
# t.test(kirc_clin$mutation_cnt ~ kirc_clin$over_surv_stt)
# 
# ggplot(kirc_clin, aes(x=over_surv_stt, y=over_surv_mth)) +
#   geom_boxplot(width = .5) +
#   geom_jitter(width = 0.05, alpha = 0.2, color = "orange")
# t.test(kirc_clin$over_surv_mth ~ kirc_clin$over_surv_stt)
# 
# ggplot(kirc_clin, aes(x=over_surv_stt, y=short_dim)) +
#   geom_boxplot(width = .5) +
#   geom_jitter(width = 0.05, alpha = 0.2, color = "orange")
# t.test(kirc_clin$short_dim ~ kirc_clin$over_surv_stt)
# 
# ggplot(kirc_clin, aes(x=over_surv_stt, y=second_long_dim)) +
#   geom_boxplot(width = .5) +
#   geom_jitter(width = 0.05, alpha = 0.2, color = "orange")
# t.test(kirc_clin$second_long_dim ~ kirc_clin$over_surv_stt)

#' 
#' Summary for continuous explanatory variables 
#' use a parametric or non-parametric test?? 
#' 
## -----------------------------------------------------------------------------

explanatory_num <- kirc_clin %>%
  select(-over_surv_stt) %>%
  select_if(is.numeric) %>%
  names
dependent <- 'over_surv_stt'

table_num <- kirc_clin %>%
  summary_factorlist(dependent, explanatory_num, p=TRUE, 
                     add_dependent_label=TRUE,  na_include = TRUE)

knitr::kable(table_num, row.names=FALSE, align=c("l", "l", "r", "r", "r"))


#' 
## -----------------------------------------------------------------------------
# Correlation Matrix
# Pearson's (normal distribution) or Spearman (not-normal) correlations
corr_num <- kirc_clin %>%
     select_if(is.numeric) %>%
     drop_na()

# Check the correlation between variables to exclude the higly correlated
cor_matrix <- cor(corr_num, method = "spearman")
cor_matrix <- round(cor_matrix, 2)

cor_matrix 

#' 
#' ## 5. Categorical variables vs. over_surv_stt
#' 
#' Tabulation and chi-square test
#' 
## ----warning=FALSE------------------------------------------------------------
# t_metas_stg <- table(kirc_clin$metastasis_stg, kirc_clin$over_surv_stt, exclude = NULL)
# t_metas_stg <- addmargins(round(100*prop.table(t_metas_stg)))
# t_metas_stg
# chisq.test(x = kirc_clin$metastasis_stg, y = kirc_clin$over_surv_stt) 

#' 
#' Summary for chategorical explanatory variables
#' Chi-squared warnings will be generated when the expected count in any cell is less than 5.
#' 
## -----------------------------------------------------------------------------
explanatory_char <- kirc_clin %>%
  select(-over_surv_stt) %>%
  select_if(is.factor) %>%
  names

dependent <- 'over_surv_stt'

table_char <- kirc_clin %>%
  summary_factorlist(dependent, explanatory_char, p=TRUE, 
                     add_dependent_label=TRUE,  na_include = TRUE)

knitr::kable(table_char, row.names=FALSE, align=c("l", "l", "r", "r", "r"))


# Droping levels with narrow distributions -> check warnings ()
# Group some levels or drop one (NULL = 'level') when grouping is not possible 

kirc_clin2 <- kirc_clin %>%
     mutate(neoplasm_stg = fct_collapse(neoplasm_stg, 'Stage I-II' = c('Stage I','Stage II'), 'Stage III-IV' = c('Stage III','Stage IV')),
            tumor_stg = fct_collapse(tumor_stg, 'T1-T2' = c('T1','T2'), 'T3-T4' = c('T3','T4')),
            histology_grd = fct_collapse(histology_grd, 'G1-G2' = c('G1','G2'), 'G3-G4' = c('G3','G4')),
            hemoglobin = fct_collapse(hemoglobin, 'Not.Low' = c('Normal', 'Elevated')),
            serum_ca = fct_collapse(serum_ca, 'Not.Low' = c('Normal', 'Elevated')),
            wbc = fct_collapse(wbc, 'Not.Elevated' = c('Low', 'Normal')))
                                                       
kirc_clin2 <- kirc_clin2 %>%
     mutate(histology_grd = fct_recode(histology_grd,  NULL = 'GX'),
            tumor_lateral = fct_recode(tumor_lateral, NULL = 'Bilateral'),
            race = fct_recode(race, NULL = 'Asian'))


table_char2 <- kirc_clin2 %>%
  summary_factorlist(dependent, explanatory_char, p=TRUE, 
                     add_dependent_label=TRUE,  na_include = TRUE)

knitr::kable(table_char2, row.names=FALSE, align=c("l", "l", "r", "r", "r"))

#' 
#' ## 6. saving dataset for regression model
#' 
## -----------------------------------------------------------------------------
kirc_glm <- kirc_clin2 
write_csv(kirc_glm, path = "data/kirc_glm.csv")

#' 
#' ## Further analysis
#' 
#' - [A logistic regression analysis](3.logistic_regression.md) of each clinical variable weight.
#' 
## -----------------------------------------------------------------------------
sessionInfo()

#'
