==> analysis/3.logistic_regression.Rmd <==
---
title: "A logistic regression analysis of TCGA-KIRC"
output: 
  html_document: 
    default
  github_document: 
    df_print: paged
    html_preview: FALSE
    keep_html: TRUE
  pdf_document:
    latex_engine: xelatex
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_format = "all") })    
---

This project contains a pipeline of clinical analysis of the Cancer Genome Atlas Kidney Renal Clear Cell Carcinoma (TCGA-KIRC) data of patients from [Genomic Data Commons Data Portal](https://portal.gdc.cancer.gov/exploration?filters=%7B%22op%22%3A%22and%22%2C%22content%22%3A%5B%7B%22op%22%3A%22in%22%2C%22content%22%3A%7B%22field%22%3A%22cases.project.project_id%22%2C%22value%22%3A%5B%22TCGA-KIRC%22%5D%7D%7D%5D%7D) and [cBioPortal](https://www.cbioportal.org/study/summary?id=kirp_tcga).

Previously, we presented a [An exploratory preprocessing analysis](1.preprocessing.md), and [a correlation analysis](2.correlation.md).

In this final section, we present a logistic regression analysis of each clinical variable weight for TCGA-KIRC.

```{r error=TRUE, message=FALSE, warning=FALSE, include=FALSE, purl=FALSE, results='hide'}
## This chunk automatically generates a text .R version of this script when running within knitr.
input  = knitr::current_input()  # filename of input document
output = paste(tools::file_path_sans_ext(input), 'R', sep = '.')
try(knitr::purl(input,output,documentation=2,quiet=T), silent=T)
# Avoid duplicate label error of knitr::purl
options(knitr.duplicate.label = 'allow')
# Code to browse the markdown file with renderized images.
knitr::opts_chunk$set(
  fig.path = "figs/3-logi_"
)
```


```{r message=FALSE, warning=FALSE, paged.print=FALSE}
# Set the packages of interest
packages = c("tidyverse","skimr","finalfit", "caret", "ggplot2", "plotROC",  "gapminder", "ggstatsplot")

# if a package is installed, it will be loaded
# otherwise, the missing package(s) will be installed and loaded
package.check <- lapply(packages, FUN = function(x) {
  if (!require(x, character.only = TRUE)) {
    install.packages(x, dependencies = TRUE)
    library(x, character.only = TRUE)
  }
})

suppressMessages(library("tidyverse"))
rm(packages)
setwd(".")
```


## 1. Importing data

```{r message=FALSE, warning=FALSE, paged.print=FALSE, echo = FALSE}

kirc_clin <- read_csv("data/kirc_clin.csv")

```



## 2. Taming data and selecting numeric and categorical columns

```{r}
kirc_clin <- kirc_clin %>%
  mutate_if(is.character, as.factor) %>%
  mutate(patient_id = as.character(patient_id),
         age = as.integer(age),
         year_diagnose = as.integer(year_diagnose))

cols_numeric <- kirc_clin %>% select_if(is.numeric) %>% names

# Removing dependent variable with high correlation: over_surv_mth, disease_free_mth and year_diagnose
cols_numeric <- setdiff(cols_numeric,  c("over_surv_mth", "disease_free_mth"))
 
cols_categorical <- kirc_clin %>% select_if(is.factor) %>% names
cols_categorical <- setdiff(cols_categorical,  "over_surv_stt") 
```



```{r}
plot(kirc_clin$over_surv_stt)
```

## 3. Creating training and test data 70-30 split

```{r}
set.seed(123)

df <- data.frame(kirc_clin)

trainIndex <- createDataPartition(df$over_surv_stt, p = .7, 
                                  list = FALSE, 
                                  times = 1)
dtrain<-df[trainIndex,]
dtest<-df[-trainIndex,]
```


## 4. Modeling: Logistic regression for numerical values with 10 fold cross validation

```{r}
fitControl <- trainControl(## 10-fold CV
  method = "cv",
  number = 10,
  savePredictions = TRUE
)

## Logistic regression
lreg<-train(x = dtrain[,c(cols_numeric)],
            y = dtrain[,"over_surv_stt"], 
            method="glm",
            family=binomial(),
            trControl=fitControl)
lreg
```
```{r}
varImp(lreg)
```


```{r}
lreg_pred<-predict(lreg,dtest)
confusionMatrix(lreg_pred,dtest$over_surv_stt)
```

## 5. Removing all nas of dataset

```{r}

set.seed(123)

kirc_clin.nona <- data.frame(kirc_clin %>% drop_na() )


trainIndex <- createDataPartition(df$over_surv_stt, p = .7, 
                                  list = FALSE, 
                                  times = 1)
dtrain<-kirc_clin.nona[trainIndex,]
dtest<-kirc_clin.nona[-trainIndex,]
```

```{r}
fitControl <- trainControl(## 10-fold CV
  method = "cv",
  number = 10,
  savePredictions = FALSE
)

## Logistic regression
lreg <- train(over_surv_stt ~.,
        data = kirc_clin.nona[ , c(cols_numeric,  cols_categorical,  "over_surv_stt")],
            method="glm",
            family=binomial(),
            trControl=fitControl)
lreg
```

```{r}
varImp(lreg)
```

```{r}
fit<-glm(over_surv_stt ~.,
        data = kirc_clin[ , c(cols_numeric, cols_categorical, "over_surv_stt")],
        family = "binomial")

varImp(fit)

```


```{r}
lreg_pred<-predict(lreg,dtest)
confusionMatrix(lreg_pred,dtest$over_surv_stt)
```

```{r}
df.nona <- dtrain %>% drop_na()

fit<-glm(over_surv_stt ~.,
        data = df.nona[ , c(cols_numeric, "over_surv_stt")],
        family = "binomial")

df_roc <- data.frame(Survival = df.nona$over_surv_stt, Prob = fit$fitted.values)
ggplot(df_roc, aes(d = Survival, m = Prob)) + geom_roc()
```


<!-- # Regression plots with finalfit -->

<!-- ```{r} -->

<!-- explanatory = c(cols_numeric, cols_categorical) -->
<!-- dependent = 'over_surv_stt' -->

<!-- kirc_clin %>% -->
<!--   or_plot(dependent, explanatory) -->

<!-- ``` -->

<!-- ```{r} -->
<!-- explanatory = c(cols_numeric, cols_categorical) -->
<!-- dependent = 'over_surv_stt' -->

<!-- kirc_clin.nona %>% -->
<!--   or_plot(dependent, explanatory) -->
<!-- ``` -->


<!-- ```{r} -->



<!-- mod <- stats::lm(formula = over_surv_stt ~ ., data = kirc_clin[ , c(cols_numeric, "over_surv_stt")]) -->

<!-- # plot -->
<!-- ggstatsplot::ggcoefstats(mod) -->

<!-- ``` -->


<!-- ```{r} -->

<!-- fit<-glm(over_surv_stt ~., -->
<!--         data = kirc_clin[ , c(cols_numeric, "over_surv_stt")], -->
<!--         family = "binomial") -->

<!-- ggstatsplot::ggcoefstats( -->
<!--   x = fit, -->
<!--   #point.args = list(color = "red", size = 3, shape = 15), -->
<!--   #vline.args = list(size = 1, color = "#CC79A7", linetype = "dotdash"), -->
<!--   #stats.label.color = c("#0072B2", "#D55E00", "darkgreen"), -->
<!--   title = "Regression model", -->
<!--   #ggtheme = hrbrthemes::theme_ipsum_ps(), -->
<!--   #ggstatsplot.layer = FALSE -->
<!-- ) #+ # note the order in which the labels are entered -->
<!--   #ggplot2::scale_y_discrete(labels = c("DECEASED", "LIVING")) + -->
<!--   #ggplot2::labs(x = "regression coefficient", y = NULL) -->

<!-- ``` -->

<!-- ```{r} -->
<!-- dependent_os = "Surv(over_surv_mth, over_surv_stt)" -->
<!-- explanatory = c("tumor_stg") -->

<!-- kirc_clin %>%  -->
<!--     surv_plot(kirc_clin.nona, explanatory, pval = TRUE) -->
<!-- #> Warning: Vectorized input to `element_text()` is not officially supported. -->
<!-- #> Results may be unexpected or may change in future versions of ggplot2. -->
<!-- ``` -->


```{r}
sessionInfo()
```
